# twoman-api/docs/swagger.yaml
definitions:
  types.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Twoman API server
  title: Twoman API
  version: "0.1"
paths:
  /friendship/accept/{profileId}:
    post:
      consumes:
      - application/json
      description: Accept a friend request
      parameters:
      - description: Friendship ID
        in: path
        name: profileId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      summary: Accept Friend Request
      tags:
      - Friendship
  /friendship/friends:
    get:
      consumes:
      - application/json
      description: Get the list of friends for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      summary: Get Friends
      tags:
      - Friendship
  /friendship/reject/{profileId}:
    post:
      consumes:
      - application/json
      description: Reject a friend request
      parameters:
      - description: Friendship ID
        in: path
        name: profileId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      summary: Reject Friend Request
      tags:
      - Friendship
  /friendship/remove/{profileId}:
    post:
      consumes:
      - application/json
      description: Remove a friendship
      parameters:
      - description: Friend's Profile ID
        in: path
        name: profileId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      summary: Remove Friendship
      tags:
      - Friendship
  /friendship/requests:
    get:
      consumes:
      - application/json
      description: Get the list of friend requests for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      summary: Get Friend Requests
      tags:
      - Friendship
  /friendship/send/{profileId}:
    post:
      consumes:
      - application/json
      description: Send a friend invite to another user
      parameters:
      - description: Recipient's User ID
        in: path
        name: profileId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      summary: Send Friend Invite
      tags:
      - Friendship
swagger: "2.0"